% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{get_datatable_past}
\alias{get_datatable_past}
\title{Interactive table of results}
\usage{
get_datatable_past(test = NULL, byte_optimize = NULL)
}
\arguments{
\item{test}{Default \code{NULL}. The default behaviour is to average over all tests.}

\item{byte_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the byte optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}
}
\description{
A summary of past results
}

