% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_set.R
\name{is_blas_optimize}
\alias{is_blas_optimize}
\title{BLAS optimize}
\usage{
is_blas_optimize(results)
}
\arguments{
\item{results}{The output from a \code{benchmark_*} call.}
}
\description{
Try to determine parallel BLAS, which implies non-standard R!
Compare user with elapsed time. If user >> elapsed, then parallel BLAS
}

