% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_set.R, R/move_files.R
\name{make_data_set}
\alias{make_data_set}
\alias{move_files}
\title{Functions for manipulating uploaded results}
\usage{
make_data_set(from)

move_files(from, to)
}
\arguments{
\item{from}{A directory containing the uploaded results.}

\item{to}{Destination directory}
}
\description{
Functions used for moving and creating the \code{past_results_v2} data set from 
uploaded data. The \code{move_files} function is used to moved files from the server
to another location, whilst removing any empty data sets.
}
\note{
One of the unit tests uploads an empty results file. 
Files where the results are \code{NULL} are moved to a sub-directory (called)
\code{empty} in the \code{to} directory. If the \code{empty} directory doesn't 
exist, it is created.

Currently these functions are specific to my set-up.
}
