% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{shine}
\alias{shine}
\title{Shiny interface}
\usage{
shine(results = NULL)
}
\arguments{
\item{results}{Benchmark results, probably obtained via the 
\code{benchmarkme} package.}
}
\description{
This function uses shiny to explore the data.
}

