% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkmeData-package.R
\docType{package}
\name{benchmarkmeData-package}
\alias{benchmarkmeData}
\alias{benchmarkmeData-package}
\title{The benchmarkmeData package}
\description{
This package contains the results from users running the 
\pkg{benchmarkme} package. The key functions are \code{plot_past()}, \code{get_datatable_past()} 
and \code{shine()}.
}
\examples{
plot_past()
## Need the DT package
get_datatable_past()
## Need the shiny package
\dontrun{
shine()
}
}
\author{
\email{csgillespie@gmail.com}
}
\seealso{
\url{https://github.com/csgillespie/benchmarkme-data}
}
\keyword{package}

