% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkmeData-package.R
\docType{package}
\name{benchmarkmeData-package}
\alias{benchmarkmeData-package}
\alias{benchmarkmeData}
\title{The benchmarkmeData package}
\description{
This package contains the results from users running the 
\pkg{benchmarkme} package. The key function is \code{plot_past()}.
}
\examples{
plot_past("prog")
}
\seealso{
\url{https://github.com/csgillespie/benchmarkme-data}
}
\author{
\email{csgillespie@gmail.com}
}
\keyword{package}
