% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{get_datatable_past}
\alias{get_datatable_past}
\title{Interactive table of results}
\usage{
get_datatable_past(test_group, byte_optimize = NULL, blas_optimize = NULL)
}
\arguments{
\item{test_group}{One of "prog", "matrix_fun", "matrix_cal", "read5", "read50", "read200", 
"write5", "write50" or "write200". Default value \code{prog}.}

\item{byte_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the byte optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}

\item{blas_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the BLAS optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}
}
\description{
A summary of past results
}
\examples{
## Need the DT package
## View all results for prog test
get_datatable_past("prog")
## View matrix_fun test
get_datatable_past("matrix_fun")
## View matrix_fun test - only BLAS results
get_datatable_past("matrix_fun", blas_optimize=TRUE)
}
