% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{get_datatable_past}
\alias{get_datatable_past}
\title{Interactive table of results}
\usage{
get_datatable_past(test_group, blas_optimize = NULL, cores = 0)
}
\arguments{
\item{test_group}{One of "prog", "matrix_fun", "matrix_cal", "read5", "read50", "read200", 
"write5", "write50" or "write200". Default value \code{prog}.}

\item{blas_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the BLAS optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}

\item{cores}{Default \code{0}, i.e. no parallel.}
}
\description{
A summary of past results
}
\examples{
## Need the DT package
## View all results for prog test
get_datatable_past("prog")
}
