% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_results.R
\name{select_results}
\alias{select_results}
\title{Selecting results}
\usage{
select_results(test_group, byte_optimize = NULL, blas_optimize = NULL)
}
\arguments{
\item{test_group}{One of "prog", "matrix_fun", "matrix_cal", "read", "write". Default it \code{prog}.}

\item{byte_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the byte optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}

\item{blas_optimize}{Default \code{NULL}. The default behaviour is to plot all results.
To plot only the BLASS optimized results, set to \code{TRUE}, otherwise \code{FALSE}.}
}
\value{
A data frame
}
\description{
Selects and  aggregates over the \code{past_results} data set.
This function just performs simple data frame subsetting.
}
\examples{
select_results("prog", byte_optimize=TRUE)
select_results("matrix_fun", blas_optimize=TRUE)
}

