% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{shine}
\alias{shine}
\title{Shiny interface}
\usage{
shine(includecode = FALSE)
}
\arguments{
\item{includecode}{Logical. If \code{TRUE} the code
included in the ui.R and server.R source files will
be shown alongside the app.}
}
\description{
This function uses shiny to explore the data.
}

