% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{shine}
\alias{shine}
\title{Shiny interface}
\usage{
shine(results = NULL)
}
\arguments{
\item{results}{Benchmark results, probably obtained via the 
\code{benchmarkme} package.}
}
\description{
This function uses the shiny package to explore past benchmarks.
The interface is launched via \code{shine()}
If you pass a benchmark, say by running \code{benchmark_std()} your
particular machine will be highlighted with a red dot.
}
\examples{
\dontrun{
shine()
## Example results
data(sample_results, package="benchmarkme")
shine(sample_results)
}
}

